requirements.txt
Django>=6.0
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
